sam/requirements.txt
adapters==1.0.1
addict==2.4.0
antlr4-python3-runtime==4.9.3
asttokens==2.4.1
attrs==23.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.8
decorator==5.1.1
executing==2.1.0
filelock==3.16.1
fonttools==4.54.1
fsspec==2024.10.0
huggingface-hub==0.26.2
hydra-core==1.3.2
idna==3.10
imageio==2.34.1
iopath==0.1.10
ipykernel==6.29.5
ipython==8.29.0
jedi==0.19.2
Jinja2==3.1.4
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mmcv==1.7.2
mmengine==0.10.5
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.4.2
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
omegaconf==2.3.0
opencv-python==4.8.0.76
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
Pillow==9.4.0
platformdirs==4.3.6
portalocker==2.10.1
prompt_toolkit==3.0.48
protobuf==5.28.3
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
regex==2024.11.6
requests==2.32.3
rich==13.9.4
safetensors==0.4.5
SAM-2 @ git+https://github.com/facebookresearch/segment-anything-2.git@c2ec8e14a185632b0a5d8b161928ceb50197eddc
sam2==0.4.1
scikit-learn==1.5.2
scipy==1.11.4
setuptools==69.5.1
six==1.16.0
stack-data==0.6.3
sympy==1.13.1
tensorboardX==2.6.2.2
termcolor==2.5.0
threadpoolctl==3.5.0
timm==0.4.12
tokenizers==0.20.3
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tornado==6.4.1
tqdm==4.67.0
traitlets==5.14.3
transformers==4.45.2
triton==3.1.0
typing_extensions==4.11.0
tzdata==2024.2
urllib3==2.2.3
wcwidth==0.2.13
wheel==0.45.0
yapf==0.43.0
